@@ -1,54 +1,54 @@
 {
-  "_args": [
-    [
-      "@angular-devkit/architect@0.13.6",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
-    ]
-  ],
-  "_development": true,
-  "_from": "@angular-devkit/architect@0.13.6",
-  "_id": "@angular-devkit/architect@0.13.6",
+  "_from": "@angular-devkit/architect@0.803.20",
+  "_id": "@angular-devkit/architect@0.803.20",
   "_inBundle": false,
-  "_integrity": "sha512-Cg9z4lmCvjt5uD00E/0tBRz3ESjYicmqT3NL/BIsNVNb+s1GwCCoPSOIM8Ss4nyGDtrdono1XKSOmkJnlzF3Cw==",
+  "_integrity": "sha512-NjyDJ61i9kh8J+qXt0E2j+P5Xsmi2mPasBzwcQyrZZGiho4zC0IFxcdxyzcsXFEupmilJKkjdt2g4QQRC5rUDQ==",
   "_location": "/@angular-devkit/architect",
-  "_phantomChildren": {
-    "tslib": "1.9.3"
-  },
+  "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@angular-devkit/architect@0.13.6",
+    "raw": "@angular-devkit/architect@0.803.20",
     "name": "@angular-devkit/architect",
     "escapedName": "@angular-devkit%2farchitect",
     "scope": "@angular-devkit",
-    "rawSpec": "0.13.6",
+    "rawSpec": "0.803.20",
     "saveSpec": null,
-    "fetchSpec": "0.13.6"
+    "fetchSpec": "0.803.20"
   },
   "_requiredBy": [
     "/@angular-devkit/build-angular",
-    "/@angular-devkit/build-webpack",
-    "/@angular/cli"
+    "/@angular-devkit/build-webpack"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular-devkit/architect/-/architect-0.13.6.tgz",
-  "_spec": "0.13.6",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/@angular-devkit/architect/-/architect-0.803.20.tgz",
+  "_shasum": "792093259d69012c0763aa9176181b255d8795c8",
+  "_spec": "@angular-devkit/architect@0.803.20",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular\\node_modules\\@angular-devkit\\build-angular",
   "author": {
     "name": "Angular Authors"
   },
   "bugs": {
     "url": "https://github.com/angular/angular-cli/issues"
   },
+  "builders": "./builders/builders.json",
+  "bundleDependencies": false,
   "dependencies": {
-    "@angular-devkit/core": "7.3.6",
-    "rxjs": "6.3.3"
+    "@angular-devkit/core": "8.3.20",
+    "rxjs": "6.4.0"
   },
+  "deprecated": false,
   "description": "Angular Build Facade",
   "engines": {
-    "node": ">= 8.9.0",
-    "npm": ">= 5.5.1"
+    "node": ">= 10.9.0",
+    "npm": ">= 6.2.0"
   },
+  "experimental": true,
   "homepage": "https://github.com/angular/angular-cli",
+  "husky": {
+    "hooks": {
+      "pre-push": "node ./bin/devkit-admin hooks/pre-push"
+    }
+  },
   "keywords": [
     "angular",
     "Angular CLI",
@@ -64,5 +64,5 @@
     "url": "git+https://github.com/angular/angular-cli.git"
   },
   "typings": "src/index.d.ts",
-  "version": "0.13.6"
+  "version": "0.803.20"
 }
