@@ -38,59 +43,78 @@
     "url": "https://github.com/angular/angular-cli/issues"
   },
   "builders": "builders.json",
+  "bundleDependencies": false,
   "dependencies": {
-    "@angular-devkit/architect": "0.13.6",
-    "@angular-devkit/build-optimizer": "0.13.6",
-    "@angular-devkit/build-webpack": "0.13.6",
-    "@angular-devkit/core": "7.3.6",
-    "@ngtools/webpack": "7.3.6",
-    "ajv": "6.9.1",
-    "autoprefixer": "9.4.6",
-    "circular-dependency-plugin": "5.0.2",
+    "@angular-devkit/architect": "0.803.20",
+    "@angular-devkit/build-optimizer": "0.803.20",
+    "@angular-devkit/build-webpack": "0.803.20",
+    "@angular-devkit/core": "8.3.20",
+    "@babel/core": "7.5.5",
+    "@babel/preset-env": "7.5.5",
+    "@ngtools/webpack": "8.3.20",
+    "ajv": "6.10.2",
+    "autoprefixer": "9.6.1",
+    "browserslist": "4.6.6",
+    "cacache": "12.0.2",
+    "caniuse-lite": "1.0.30000989",
+    "circular-dependency-plugin": "5.2.0",
     "clean-css": "4.2.1",
-    "copy-webpack-plugin": "4.6.0",
-    "file-loader": "3.0.1",
-    "glob": "7.1.3",
+    "copy-webpack-plugin": "5.0.4",
+    "core-js": "3.2.1",
+    "file-loader": "4.2.0",
+    "find-cache-dir": "3.0.0",
+    "glob": "7.1.4",
     "istanbul-instrumenter-loader": "3.0.1",
-    "karma-source-map-support": "1.3.0",
+    "jest-worker": "24.9.0",
+    "karma-source-map-support": "1.4.0",
     "less": "3.9.0",
-    "less-loader": "4.1.0",
-    "license-webpack-plugin": "2.1.0",
+    "less-loader": "5.0.0",
+    "license-webpack-plugin": "2.1.2",
     "loader-utils": "1.2.3",
-    "mini-css-extract-plugin": "0.5.0",
+    "mini-css-extract-plugin": "0.8.0",
     "minimatch": "3.0.4",
-    "node-sass": "4.11.0",
-    "opn": "5.4.0",
+    "open": "6.4.0",
     "parse5": "4.0.0",
-    "postcss": "7.0.14",
+    "postcss": "7.0.17",
     "postcss-import": "12.0.1",
     "postcss-loader": "3.0.0",
-    "raw-loader": "1.0.0",
-    "rxjs": "6.3.3",
-    "sass-loader": "7.1.0",
-    "semver": "5.6.0",
+    "raw-loader": "3.1.0",
+    "regenerator-runtime": "0.13.3",
+    "rxjs": "6.4.0",
+    "sass": "1.22.9",
+    "sass-loader": "7.2.0",
+    "semver": "6.3.0",
+    "source-map": "0.7.3",
     "source-map-loader": "0.2.4",
-    "source-map-support": "0.5.10",
-    "speed-measure-webpack-plugin": "1.3.0",
-    "stats-webpack-plugin": "0.7.0",
-    "style-loader": "0.23.1",
+    "source-map-support": "0.5.13",
+    "speed-measure-webpack-plugin": "1.3.1",
+    "style-loader": "1.0.0",
     "stylus": "0.54.5",
     "stylus-loader": "3.0.2",
-    "terser-webpack-plugin": "1.2.2",
+    "terser": "4.3.9",
+    "terser-webpack-plugin": "1.4.1",
     "tree-kill": "1.2.1",
-    "webpack": "4.29.0",
-    "webpack-dev-middleware": "3.5.1",
-    "webpack-dev-server": "3.1.14",
+    "webpack": "4.39.2",
+    "webpack-dev-middleware": "3.7.2",
+    "webpack-dev-server": "3.9.0",
     "webpack-merge": "4.2.1",
-    "webpack-sources": "1.3.0",
-    "webpack-subresource-integrity": "1.1.0-rc.6"
+    "webpack-sources": "1.4.3",
+    "webpack-subresource-integrity": "1.1.0-rc.6",
+    "worker-plugin": "3.2.0"
   },
+  "deprecated": false,
   "description": "Angular Webpack Build Facade",
   "engines": {
-    "node": ">= 8.9.0",
-    "npm": ">= 5.5.1"
+    "node": ">= 10.9.0",
+    "npm": ">= 6.2.0"
   },
+  "experimental": true,
   "homepage": "https://github.com/angular/angular-cli",
+  "husky": {
+    "hooks": {
+      "pre-push": "node ./bin/devkit-admin hooks/pre-push"
+    }
+  },
   "keywords": [
     "angular",
     "Angular CLI",
