@@ -1,34 +1,28 @@
 {
-  "_args": [
-    [
-      "@angular-devkit/build-optimizer@0.13.6",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
-    ]
-  ],
-  "_development": true,
-  "_from": "@angular-devkit/build-optimizer@0.13.6",
-  "_id": "@angular-devkit/build-optimizer@0.13.6",
+  "_from": "@angular-devkit/build-optimizer@0.803.20",
+  "_id": "@angular-devkit/build-optimizer@0.803.20",
   "_inBundle": false,
-  "_integrity": "sha512-/EOO0wxw7FIKfove4Fv/aKiKgvH0kFhRT2GOe3pW0yV2y2e1dfQdo2AqdGCfJ4rSbwsfWSGTUUFITc+9jgXFZw==",
+  "_integrity": "sha512-Vzxf1g1EuzaPBoScDYUhyxemi5chlgnpWmObNo5dzVAVzjxo5gJeDIGpiyDqHvr6LBkprqb6XHcZhMWqIcdIHg==",
   "_location": "/@angular-devkit/build-optimizer",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@angular-devkit/build-optimizer@0.13.6",
+    "raw": "@angular-devkit/build-optimizer@0.803.20",
     "name": "@angular-devkit/build-optimizer",
     "escapedName": "@angular-devkit%2fbuild-optimizer",
     "scope": "@angular-devkit",
-    "rawSpec": "0.13.6",
+    "rawSpec": "0.803.20",
     "saveSpec": null,
-    "fetchSpec": "0.13.6"
+    "fetchSpec": "0.803.20"
   },
   "_requiredBy": [
     "/@angular-devkit/build-angular"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular-devkit/build-optimizer/-/build-optimizer-0.13.6.tgz",
-  "_spec": "0.13.6",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/@angular-devkit/build-optimizer/-/build-optimizer-0.803.20.tgz",
+  "_shasum": "36f92c350c1cb3d3d7f1e0f0086feed726478902",
+  "_spec": "@angular-devkit/build-optimizer@0.803.20",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular\\node_modules\\@angular-devkit\\build-angular",
   "author": {
     "name": "Angular Authors"
   },
@@ -38,18 +32,27 @@
   "bugs": {
     "url": "https://github.com/angular/angular-cli/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
     "loader-utils": "1.2.3",
-    "source-map": "0.5.6",
-    "typescript": "3.2.4",
-    "webpack-sources": "1.3.0"
+    "source-map": "0.7.3",
+    "tslib": "1.10.0",
+    "typescript": "3.5.3",
+    "webpack-sources": "1.4.3"
   },
+  "deprecated": false,
   "description": "Angular Build Optimizer",
   "engines": {
-    "node": ">= 8.9.0",
-    "npm": ">= 5.5.1"
+    "node": ">= 10.9.0",
+    "npm": ">= 6.2.0"
   },
+  "experimental": true,
   "homepage": "https://github.com/angular/angular-cli",
+  "husky": {
+    "hooks": {
+      "pre-push": "node ./bin/devkit-admin hooks/pre-push"
+    }
+  },
   "keywords": [
     "angular",
     "Angular CLI",
@@ -65,5 +68,5 @@
     "url": "git+https://github.com/angular/angular-cli.git"
   },
   "typings": "src/index.d.ts",
-  "version": "0.13.6"
+  "version": "0.803.20"
 }
