@@ -1,36 +1,28 @@
 {
-  "_args": [
-    [
-      "@angular-devkit/build-webpack@0.13.6",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
-    ]
-  ],
-  "_development": true,
-  "_from": "@angular-devkit/build-webpack@0.13.6",
-  "_id": "@angular-devkit/build-webpack@0.13.6",
+  "_from": "@angular-devkit/build-webpack@0.803.20",
+  "_id": "@angular-devkit/build-webpack@0.803.20",
   "_inBundle": false,
-  "_integrity": "sha512-lQmYrhOHeBn/r7uw+FfpJUq9puApYe72+mrIfg/UtLwSEYt58n5d156jnuMcfWnCyhYUgYqLTaOUfSM0RV7akw==",
+  "_integrity": "sha512-35af8kD3KG/cIv7AB09YNER5HIPlx55ipBxdVk8D+X3MuUcTmD6fFvqXcV0EPlD1vQephthfzSgtNpvuPv4xuA==",
   "_location": "/@angular-devkit/build-webpack",
-  "_phantomChildren": {
-    "tslib": "1.9.3"
-  },
+  "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@angular-devkit/build-webpack@0.13.6",
+    "raw": "@angular-devkit/build-webpack@0.803.20",
     "name": "@angular-devkit/build-webpack",
     "escapedName": "@angular-devkit%2fbuild-webpack",
     "scope": "@angular-devkit",
-    "rawSpec": "0.13.6",
+    "rawSpec": "0.803.20",
     "saveSpec": null,
-    "fetchSpec": "0.13.6"
+    "fetchSpec": "0.803.20"
   },
   "_requiredBy": [
     "/@angular-devkit/build-angular"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular-devkit/build-webpack/-/build-webpack-0.13.6.tgz",
-  "_spec": "0.13.6",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/@angular-devkit/build-webpack/-/build-webpack-0.803.20.tgz",
+  "_shasum": "c03bd23a9f147c07201db681bf9fc3bfad95f9d9",
+  "_spec": "@angular-devkit/build-webpack@0.803.20",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular\\node_modules\\@angular-devkit\\build-angular",
   "author": {
     "name": "Angular Authors"
   },
@@ -38,17 +30,25 @@
     "url": "https://github.com/angular/angular-cli/issues"
   },
   "builders": "builders.json",
+  "bundleDependencies": false,
   "dependencies": {
-    "@angular-devkit/architect": "0.13.6",
-    "@angular-devkit/core": "7.3.6",
-    "rxjs": "6.3.3"
+    "@angular-devkit/architect": "0.803.20",
+    "@angular-devkit/core": "8.3.20",
+    "rxjs": "6.4.0"
   },
+  "deprecated": false,
   "description": "Webpack Builder for Architect",
   "engines": {
-    "node": ">= 8.9.0",
-    "npm": ">= 5.5.1"
+    "node": ">= 10.9.0",
+    "npm": ">= 6.2.0"
   },
+  "experimental": true,
   "homepage": "https://github.com/angular/angular-cli",
+  "husky": {
+    "hooks": {
+      "pre-push": "node ./bin/devkit-admin hooks/pre-push"
+    }
+  },
   "keywords": [
     "angular",
     "Angular CLI",
@@ -68,5 +68,5 @@
     "url": "git+https://github.com/angular/angular-cli.git"
   },
   "typings": "src/index.d.ts",
-  "version": "0.13.6"
+  "version": "0.803.20"
 }
