@@ -1,62 +1,63 @@
 {
-  "_args": [
-    [
-      "@angular-devkit/core@7.3.6",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
-    ]
-  ],
-  "_development": true,
-  "_from": "@angular-devkit/core@7.3.6",
-  "_id": "@angular-devkit/core@7.3.6",
+  "_from": "@angular-devkit/core@8.3.20",
+  "_id": "@angular-devkit/core@8.3.20",
   "_inBundle": false,
-  "_integrity": "sha512-aoarMK0DJIdwjVA0OuQIN7b8nKPcF9n5vSMF7MFmhKpTw5/uV3SynQZbm3YCgylu/2CMuiTzKuAunnWWdli//g==",
+  "_integrity": "sha512-UCfW/BJBJnioJU34QennQhA4o+rLoCXWiSrI2LM7yw8/MEM9I8KbqRETP1My3HjHkQnvP+Qh3noedpcu3Nnt8A==",
   "_location": "/@angular-devkit/core",
   "_phantomChildren": {
-    "tslib": "1.9.3"
+    "fast-deep-equal": "2.0.1",
+    "fast-json-stable-stringify": "2.0.0",
+    "json-schema-traverse": "0.4.1",
+    "sourcemap-codec": "1.4.4",
+    "uri-js": "4.2.2"
   },
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@angular-devkit/core@7.3.6",
+    "raw": "@angular-devkit/core@8.3.20",
     "name": "@angular-devkit/core",
     "escapedName": "@angular-devkit%2fcore",
     "scope": "@angular-devkit",
-    "rawSpec": "7.3.6",
+    "rawSpec": "8.3.20",
     "saveSpec": null,
-    "fetchSpec": "7.3.6"
+    "fetchSpec": "8.3.20"
   },
   "_requiredBy": [
     "/@angular-devkit/architect",
     "/@angular-devkit/build-angular",
     "/@angular-devkit/build-webpack",
-    "/@angular-devkit/schematics",
-    "/@angular/cli",
-    "/@ngtools/webpack",
-    "/@schematics/angular",
-    "/@schematics/update"
+    "/@ngtools/webpack"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular-devkit/core/-/core-7.3.6.tgz",
-  "_spec": "7.3.6",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/@angular-devkit/core/-/core-8.3.20.tgz",
+  "_shasum": "8a9e17e706905efab9875c16690e1d9328c9d5dd",
+  "_spec": "@angular-devkit/core@8.3.20",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular\\node_modules\\@angular-devkit\\build-angular",
   "author": {
     "name": "Angular Authors"
   },
   "bugs": {
     "url": "https://github.com/angular/angular-cli/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
-    "ajv": "6.9.1",
-    "chokidar": "2.0.4",
+    "ajv": "6.10.2",
     "fast-json-stable-stringify": "2.0.0",
-    "rxjs": "6.3.3",
+    "magic-string": "0.25.3",
+    "rxjs": "6.4.0",
     "source-map": "0.7.3"
   },
+  "deprecated": false,
   "description": "Angular DevKit - Core Utility Library",
   "engines": {
-    "node": ">= 8.9.0",
-    "npm": ">= 5.5.1"
+    "node": ">= 10.9.0",
+    "npm": ">= 6.2.0"
   },
   "homepage": "https://github.com/angular/angular-cli",
+  "husky": {
+    "hooks": {
+      "pre-push": "node ./bin/devkit-admin hooks/pre-push"
+    }
+  },
   "keywords": [
     "angular",
     "Angular CLI",
@@ -73,5 +74,5 @@
     "url": "git+https://github.com/angular/angular-cli.git"
   },
   "typings": "src/index.d.ts",
-  "version": "7.3.6"
+  "version": "8.3.20"
 }
