@@ -1,38 +1,44 @@
 {
   "_args": [
     [
-      "@angular-devkit/schematics@7.3.6",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
+      "@angular-devkit/schematics@8.3.20",
+      "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular"
     ]
   ],
   "_development": true,
-  "_from": "@angular-devkit/schematics@7.3.6",
-  "_id": "@angular-devkit/schematics@7.3.6",
+  "_from": "@angular-devkit/schematics@8.3.20",
+  "_id": "@angular-devkit/schematics@8.3.20",
   "_inBundle": false,
-  "_integrity": "sha512-YXF7QusmMy3D9H0vNczc1n5BkuEHLwt7cW33euNeGNgTIsD0n6DrUhgClurXicnr2GNPSDYE5+3115lmJkhyrg==",
+  "_integrity": "sha512-sDHZakh4e3A5WenR9zr1x6Va9GNRqQlRhqT3xcbkG88v2M0YqEt7dHB7YwnOhm7zSxiWQM8PdWEQHiQ4iu9NyQ==",
   "_location": "/@angular-devkit/schematics",
   "_phantomChildren": {
-    "tslib": "1.9.3"
+    "fast-deep-equal": "2.0.1",
+    "fast-json-stable-stringify": "2.0.0",
+    "json-schema-traverse": "0.4.1",
+    "rxjs": "6.4.0",
+    "source-map": "0.7.3",
+    "sourcemap-codec": "1.4.4",
+    "uri-js": "4.2.2"
   },
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@angular-devkit/schematics@7.3.6",
+    "raw": "@angular-devkit/schematics@8.3.20",
     "name": "@angular-devkit/schematics",
     "escapedName": "@angular-devkit%2fschematics",
     "scope": "@angular-devkit",
-    "rawSpec": "7.3.6",
+    "rawSpec": "8.3.20",
     "saveSpec": null,
-    "fetchSpec": "7.3.6"
+    "fetchSpec": "8.3.20"
   },
   "_requiredBy": [
     "/@angular/cli",
     "/@schematics/angular",
     "/@schematics/update"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular-devkit/schematics/-/schematics-7.3.6.tgz",
-  "_spec": "7.3.6",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/@angular-devkit/schematics/-/schematics-8.3.20.tgz",
+  "_spec": "8.3.20",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular",
   "author": {
     "name": "Angular Authors"
   },
@@ -40,15 +46,20 @@
     "url": "https://github.com/angular/angular-cli/issues"
   },
   "dependencies": {
-    "@angular-devkit/core": "7.3.6",
-    "rxjs": "6.3.3"
+    "@angular-devkit/core": "8.3.20",
+    "rxjs": "6.4.0"
   },
   "description": "Angular Schematics - Library",
   "engines": {
-    "node": ">= 8.9.0",
-    "npm": ">= 5.5.1"
+    "node": ">= 10.9.0",
+    "npm": ">= 6.2.0"
   },
   "homepage": "https://github.com/angular/angular-cli",
+  "husky": {
+    "hooks": {
+      "pre-push": "node ./bin/devkit-admin hooks/pre-push"
+    }
+  },
   "keywords": [
     "angular",
     "Angular CLI",
@@ -70,5 +81,5 @@
     "url": "git+https://github.com/angular/angular-cli.git"
   },
   "typings": "src/index.d.ts",
-  "version": "7.3.6"
+  "version": "8.3.20"
 }
