@@ -1,33 +1,33 @@
 {
   "_args": [
     [
-      "@angular/animations@7.2.9",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
+      "@angular/animations@8.2.14",
+      "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular"
     ]
   ],
-  "_from": "@angular/animations@7.2.9",
-  "_id": "@angular/animations@7.2.9",
+  "_from": "@angular/animations@8.2.14",
+  "_id": "@angular/animations@8.2.14",
   "_inBundle": false,
-  "_integrity": "sha512-cdIf/UWwhNlw6Ep3pRxRk12F8q/GyTbcHDF4qbgg7lZEDPYS32KS1lpHY1pgSuSOIPZUT+iLPOxAZ6cu7Am1vA==",
+  "_integrity": "sha512-3Vc9TnNpKdtvKIXcWDFINSsnwgEMiDmLzjceWg1iYKwpeZGQahUXPoesLwQazBMmxJzQiA4HOMj0TTXKZ+Jzkg==",
   "_location": "/@angular/animations",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@angular/animations@7.2.9",
+    "raw": "@angular/animations@8.2.14",
     "name": "@angular/animations",
     "escapedName": "@angular%2fanimations",
     "scope": "@angular",
-    "rawSpec": "7.2.9",
+    "rawSpec": "8.2.14",
     "saveSpec": null,
-    "fetchSpec": "7.2.9"
+    "fetchSpec": "8.2.14"
   },
   "_requiredBy": [
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/animations/-/animations-7.2.9.tgz",
-  "_spec": "7.2.9",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/@angular/animations/-/animations-8.2.14.tgz",
+  "_spec": "8.2.14",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular",
   "author": {
     "name": "angular"
   },
@@ -60,7 +60,6 @@
       "@angular/platform-browser",
       "@angular/platform-browser-dynamic",
       "@angular/forms",
-      "@angular/http",
       "@angular/platform-server",
       "@angular/platform-webworker",
       "@angular/platform-webworker-dynamic",
@@ -71,7 +70,10 @@
     ]
   },
   "peerDependencies": {
-    "@angular/core": "7.2.9"
+    "@angular/core": "8.2.14"
+  },
+  "publishConfig": {
+    "registry": "https://wombat-dressing-room.appspot.com"
   },
   "repository": {
     "type": "git",
@@ -79,5 +81,5 @@
   },
   "sideEffects": false,
   "typings": "./animations.d.ts",
-  "version": "7.2.9"
+  "version": "8.2.14"
 }
