@@ -39,26 +48,38 @@
     "url": "https://github.com/angular/angular-cli/issues"
   },
   "dependencies": {
-    "@angular-devkit/architect": "0.13.6",
-    "@angular-devkit/core": "7.3.6",
-    "@angular-devkit/schematics": "7.3.6",
-    "@schematics/angular": "7.3.6",
-    "@schematics/update": "0.13.6",
+    "@angular-devkit/architect": "0.803.20",
+    "@angular-devkit/core": "8.3.20",
+    "@angular-devkit/schematics": "8.3.20",
+    "@schematics/angular": "8.3.20",
+    "@schematics/update": "0.803.20",
     "@yarnpkg/lockfile": "1.1.0",
+    "ansi-colors": "4.1.1",
+    "debug": "^4.1.1",
     "ini": "1.3.5",
-    "inquirer": "6.2.1",
+    "inquirer": "6.5.1",
     "npm-package-arg": "6.1.0",
-    "opn": "5.4.0",
-    "pacote": "9.4.0",
-    "semver": "5.6.0",
-    "symbol-observable": "1.2.0"
+    "npm-pick-manifest": "3.0.2",
+    "open": "6.4.0",
+    "pacote": "9.5.5",
+    "read-package-tree": "5.3.1",
+    "rimraf": "3.0.0",
+    "semver": "6.3.0",
+    "symbol-observable": "1.2.0",
+    "universal-analytics": "^0.4.20",
+    "uuid": "^3.3.2"
   },
   "description": "CLI tool for Angular",
   "engines": {
-    "node": ">= 8.9.0",
-    "npm": ">= 5.5.1"
+    "node": ">= 10.9.0",
+    "npm": ">= 6.2.0"
   },
   "homepage": "https://github.com/angular/angular-cli",
+  "husky": {
+    "hooks": {
+      "pre-push": "node ./bin/devkit-admin hooks/pre-push"
+    }
+  },
   "keywords": [
     "angular",
     "Angular CLI",
