@@ -1,33 +1,33 @@
 {
   "_args": [
     [
-      "@angular/common@7.2.9",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
+      "@angular/common@8.2.14",
+      "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular"
     ]
   ],
-  "_from": "@angular/common@7.2.9",
-  "_id": "@angular/common@7.2.9",
+  "_from": "@angular/common@8.2.14",
+  "_id": "@angular/common@8.2.14",
   "_inBundle": false,
-  "_integrity": "sha512-SBIinZi/R0mKWjbKvcBhwRpAVIY/7/AnzIEkMb5JnGExersYhDD02otW3ZwPStpl7HrhgWk8ueWnxVFFXRz5gA==",
+  "_integrity": "sha512-Qmt+aX2quUW54kaNT7QH7WGXnFxr/cC2C6sf5SW5SdkZfDQSiz8IaItvieZfXVQUbBOQKFRJ7TlSkt0jI/yjvw==",
   "_location": "/@angular/common",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@angular/common@7.2.9",
+    "raw": "@angular/common@8.2.14",
     "name": "@angular/common",
     "escapedName": "@angular%2fcommon",
     "scope": "@angular",
-    "rawSpec": "7.2.9",
+    "rawSpec": "8.2.14",
     "saveSpec": null,
-    "fetchSpec": "7.2.9"
+    "fetchSpec": "8.2.14"
   },
   "_requiredBy": [
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/common/-/common-7.2.9.tgz",
-  "_spec": "7.2.9",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/@angular/common/-/common-8.2.14.tgz",
+  "_spec": "8.2.14",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular",
   "author": {
     "name": "angular"
   },
@@ -61,7 +61,6 @@
       "@angular/platform-browser",
       "@angular/platform-browser-dynamic",
       "@angular/forms",
-      "@angular/http",
       "@angular/platform-server",
       "@angular/platform-webworker",
       "@angular/platform-webworker-dynamic",
@@ -72,8 +71,11 @@
     ]
   },
   "peerDependencies": {
-    "rxjs": "^6.0.0",
-    "@angular/core": "7.2.9"
+    "rxjs": "^6.4.0",
+    "@angular/core": "8.2.14"
+  },
+  "publishConfig": {
+    "registry": "https://wombat-dressing-room.appspot.com"
   },
   "repository": {
     "type": "git",
@@ -81,5 +83,5 @@
   },
   "sideEffects": false,
   "typings": "./common.d.ts",
-  "version": "7.2.9"
+  "version": "8.2.14"
 }
