@@ -1,34 +1,33 @@
 {
   "_args": [
     [
-      "camelcase@4.1.0",
-      "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend"
+      "camelcase@5.3.1",
+      "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular"
     ]
   ],
   "_development": true,
-  "_from": "camelcase@4.1.0",
-  "_id": "camelcase@4.1.0",
+  "_from": "camelcase@5.3.1",
+  "_id": "camelcase@5.3.1",
   "_inBundle": false,
-  "_integrity": "sha1-1UVjW+HjPFQmScaRc+Xeas+uNN0=",
+  "_integrity": "sha512-L28STB170nwWS63UjtlEOE3dldQApaJXZkOI1uMFfzf3rRuPegHaHesyee+YxQ+W6SvRDQV6UrdOdRiR153wJg==",
   "_location": "/@angular/compiler-cli/camelcase",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "camelcase@4.1.0",
+    "raw": "camelcase@5.3.1",
     "name": "camelcase",
     "escapedName": "camelcase",
-    "rawSpec": "4.1.0",
+    "rawSpec": "5.3.1",
     "saveSpec": null,
-    "fetchSpec": "4.1.0"
+    "fetchSpec": "5.3.1"
   },
   "_requiredBy": [
-    "/@angular/compiler-cli/yargs",
     "/@angular/compiler-cli/yargs-parser"
   ],
-  "_resolved": "https://registry.npmjs.org/camelcase/-/camelcase-4.1.0.tgz",
-  "_spec": "4.1.0",
-  "_where": "M:\\UNIVERSIDAD\\SEGUNDO CUATRIMESTRE\\DAW\\Repo\\santatecla-relaciones-2\\frontend",
+  "_resolved": "https://registry.npmjs.org/camelcase/-/camelcase-5.3.1.tgz",
+  "_spec": "5.3.1",
+  "_where": "M:\\UNIVERSIDAD\\TFG\\2019-LicenseManager\\angular",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -37,16 +36,18 @@
   "bugs": {
     "url": "https://github.com/sindresorhus/camelcase/issues"
   },
-  "description": "Convert a dash/dot/underscore/space separated string to camelCase: foo-bar → fooBar",
+  "description": "Convert a dash/dot/underscore/space separated string to camelCase or PascalCase: `foo-bar` → `fooBar`",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^1.4.1",
+    "tsd": "^0.7.1",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/sindresorhus/camelcase#readme",
   "keywords": [
@@ -61,7 +62,9 @@
     "separator",
     "string",
     "text",
-    "convert"
+    "convert",
+    "pascalcase",
+    "pascal-case"
   ],
   "license": "MIT",
   "name": "camelcase",
@@ -70,10 +73,7 @@
     "url": "git+https://github.com/sindresorhus/camelcase.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd"
   },
-  "version": "4.1.0",
-  "xo": {
-    "esnext": true
-  }
+  "version": "5.3.1"
 }
