@@ -1,9 +1,11 @@
-import { Component, ChangeDetectorRef, AfterViewInit, ViewChild, TemplateRef } from '@angular/core';
+import { Component, ChangeDetectorRef, AfterViewInit, ViewChild, TemplateRef, OnDestroy, ElementRef, OnInit } from '@angular/core';
 import { environment } from 'src/environments/environment';
 import { TdTabSelectBase } from '@covalent/core';
 import { Router } from '@angular/router';
+import { NgForm } from '@angular/forms';
+import { UserProfileService } from './userProfile/userProfile.service';
 
-
+declare var Stripe: any;
 @Component({
     selector: 'my-app',
     templateUrl: './app.component.html',
@@ -13,7 +15,9 @@ import { Router } from '@angular/router';
 export class AppComponent {
     mode: string;
 
-    constructor(private router: Router) {
+    
+    
+    constructor(private router: Router,private userProfileServ:UserProfileService) {
         if (environment.production) {
             this.mode = "Production";
         } else {
@@ -21,6 +25,7 @@ export class AppComponent {
         }
     }
 
+      
     manageLinks(route:string){
         switch(route){
             case 'adminDashboard':{
