@@ -17,40 +17,14 @@ export class UserProfileService {
         
     }
 
-    addCard(owner:string,cvv:string,number:string,exp_month:string,exp_year:string){
-        let params = new Map();
-        params.set("owner",owner);
-        params.set("cvv",cvv);
-        params.set("exp_month",exp_month);
-        params.set("exp_year",exp_year);
-        params.set("number",number);
 
-        let card = {owner:owner,cvv:cvv,exp_month:exp_month,exp_year:exp_year,number:number};
-
-        
-        //let params:string[]= [number,cvv,exp_month,exp_year];
-
-        console.log(card);
-
-        return this.http.post(BASE_URL + this.loginServ.user.name + '/addCard',card).pipe(
+    addCardStripeElements(tokenId:string){
+        return this.http.post(BASE_URL + 'addCard/' + tokenId,null).pipe(
             map(response => console.log(response)),
         );
     }
 
-    
-    updateCards(){
-        return this.http.put(BASE_URL,this.loginServ.user)
-            .map(response => response.json())
-		    .catch(error => this.handleError(error));
-       
-    }
 
-
-    getCards(){
-        return this.http.get(BASE_URL+"stripeCards")
-        .map(response => response.json())
-        .catch(error => this.handleError(error));
-    }
 
     addSubscriptionToProduct(product:Product,typeSubs:string,userName:string){
         return this.http.put(BASE_URL+product.name+"/" + typeSubs + "/" + userName +"/addSubscription",product)
